labextension: Add a KFServing dialog
This commit also introduces two form components: FormInput and
FormSelect. We need them because using react-hook-form with uncontrolled
components (that is, we don't manually apply a value and react to its
change) is much easier.

NOTE: FormSelect is almost identical to Select. Consider reimplementing
FormSelect as a class component and have Select inherit from it.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>